MUI package and updates to main APP component
